Move scale function to helpers file
